refactored services directory
