+++ frontend/src/pages/account/EditProfile/Content/components/PrivacySettings/Form/onSubmit.ts
@@ -1,8 +1,8 @@
 import { UseFormSetError } from 'react-hook-form'
 import { Dispatch, SetStateAction } from 'react'
 
-import EditUserService from 'services/editUserService'
 import { PrivacySettingsForm } from './interfaces'
+import { EditUserService } from 'services'
 
 const onSubmit = async (
   data: PrivacySettingsForm,
