+++ frontend/src/redux/slices/myProfile/thunks/logout.ts
@@ -1,6 +1,5 @@
 import { createAsyncThunk } from '@reduxjs/toolkit'
-
-import AuthService from 'services/authService'
+import { AuthService } from 'services'
 
 const logout = createAsyncThunk(
   'myProfile/logout',
