+++ frontend/src/services/myUserService.ts
@@ -3,7 +3,7 @@ import { AxiosResponse } from 'axios'
 import $api from 'http/axios'
 import { myUser } from 'models'
 
-export default class MyUserService {
+export class MyUserService {
   static fetchMyData(): Promise<AxiosResponse<myUser.IMyUser>> {
     return $api.get<myUser.IMyUser>('/user')
   }
